fix :: dynamic canvas background
