+++ frontend/src/_ui/FolderList/FolderList.scss
@@ -1,5 +1,4 @@
 .tj-list-item {
-    font-family: 'Roboto';
     font-style: normal;
     font-weight: 400;
     font-size: 12px;
